// libs/spriced-common-lib/src/lib/services/dynamic-form.types.ts
export interface QueryColumns {
  name: string;
  displayName?: string;
  dataType: "number" | "string" | "date" | "boolean" | "category";
  options?: { name: string; value: any }[];
  nullable?: boolean;
}

export const FORM_DATA_SERVICE = 'FORM_DATA_SERVICE';


export interface Header {
  name: string;
  column: string;
  width?: number;
  hidden?: boolean;
  pinned?: "left" | "right";
  flexGrow?: number;
  isSortable?: boolean;
  isFilterable?: boolean;
  sortDirection?: "asc" | "desc";
  canAutoResize?: boolean;
  pipe?: unknown;
  //contentProjection?: boolean;
  checkbox?: boolean;
  tooltip?: boolean;
  tooltipTemplate?: any;
  imgsrc?: any;
  disableCheckbox?: any;
  isLink?: boolean;
}